Moves function into cluster reconciler
